Added Hystrix circuit breaker
The Hystrix work as a proxy class to the class enabled with annotation @EnableCircuitBreaker.
So if the method calling the method with fallbacks and fallback methods are in the same class then the fallbacks don't get executed, because the proxy class is one level above. For the fallbacks to get executed, they need to be in seperate class.
For this purpose, two classes called MovieInfo and UserRatingInfo are created and fallback methods are moved to it.